exception handling delete

+++ src/main/java/com/delivery/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.delivery.course.entities.User;
 import com.delivery.course.repositories.UserRepository;
+import com.delivery.course.services.exceptions.DatabaseException;
 import com.delivery.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -32,7 +34,16 @@ public class UserService {
 	}
 
 	public void delete(Long id) {
-		repository.deleteById(id);
+		Optional<User> user = repository.findById(id);
+		if (user.isPresent()) {
+			try {
+				repository.deleteById(id);
+			} catch (DataIntegrityViolationException e) {
+				throw new DatabaseException(e.getMessage());
+			}
+		} else {
+			throw new ResourceNotFoundException(id);
+		}
 	}
 
 	public User update(Long id, User obj) {
